--- app/src/main/java/com/example/ihonboarding/data/data_source/remote/NewsRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.example.ihonboarding.data.data_source.remote
-
-import com.example.ihonboarding.data.model.NewsList
-import retrofit2.Response
-
-interface NewsRemoteDataSource {
-    suspend fun getNews(): Response<NewsList>
-}
