+++ app/src/main/java/com/example/ihonboarding/data/home/data_source/remote/NewsRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.example.ihonboarding.data.home.data_source.remote
+
+import com.example.ihonboarding.data.home.model.NewsList
+import retrofit2.Response
+
+interface NewsRemoteDataSource {
+    suspend fun getNews(): Response<NewsList>
+}
