+++ app/src/main/java/com/example/ihonboarding/data/home/db/NewsDb.kt
@@ -1,0 +1,11 @@
+package com.example.ihonboarding.data.home.db
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import com.example.ihonboarding.data.home.dao.NewsDao
+import com.example.ihonboarding.data.home.model.News
+
+@Database(entities = [News::class], version = 1, exportSchema = false)
+abstract class NewsDb : RoomDatabase() {
+    abstract fun newsDao(): NewsDao
+}
