+++ app/src/main/java/com/example/ihonboarding/data/home/model/Counter.kt
@@ -1,0 +1,20 @@
+package com.example.ihonboarding.data.home.model
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Counter(
+    @SerialName("comment")
+    val comment: Int? = 0,
+
+    @SerialName("downvote")
+    val downvote: Int? = 0,
+
+    @SerialName("upvote")
+    val upvote: Int? = 0,
+
+    @SerialName("view")
+    val view: Int? = 0
+)
