+++ app/src/main/java/com/example/ihonboarding/data/home/model/NewsList.kt
@@ -1,0 +1,11 @@
+package com.example.ihonboarding.data.home.model
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class NewsList(
+    @SerialName("results")
+    val news: List<News>? = listOf()
+)
