--- app/src/main/java/com/example/ihonboarding/data/model/NewsList.kt
@@ -1,11 +1,0 @@
-package com.example.ihonboarding.data.model
-
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class NewsList(
-    @SerialName("results")
-    val news: List<News>? = listOf()
-)
