--- app/src/main/java/com/example/ihonboarding/domain/repository/NewsRepository.kt
@@ -1,8 +1,0 @@
-package com.example.ihonboarding.domain.repository
-
-import com.example.ihonboarding.data.model.News
-
-interface NewsRepository {
-    suspend fun getNews(): List<News>?
-    suspend fun updateNews(): List<News>?
-}
