+++ app/src/main/java/com/example/ihonboarding/presentation/ui/home/BaseApplication.kt
@@ -1,0 +1,8 @@
+package com.example.ihonboarding.presentation.ui.home
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class BaseApplication : Application() {
+}
