+++ app/src/main/java/com/example/ihonboarding/presentation/ui/home/NewsFragment.kt
@@ -1,0 +1,50 @@
+package com.example.ihonboarding.presentation.ui.home
+
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.platform.ComposeView
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import com.example.ihonboarding.presentation.screen.home.HomePage
+import com.example.ihonboarding.presentation.theme.IHOnboardingTheme
+import com.example.ihonboarding.presentation.viewmodel.NewsListViewModel
+import com.example.ihonboarding.util.Constants.Companion.TAG
+import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class NewsFragment : Fragment() {
+    @Inject
+    lateinit var app: BaseApplication
+
+    private val viewModel: NewsListViewModel by viewModels()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+
+        return ComposeView(requireContext()).apply {
+            setContent {
+
+                val news = viewModel.news.value
+
+                for(item in news) {
+                    Log.d(TAG, "onCreateView:${item.title}")
+                }
+
+                IHOnboardingTheme {
+                    Surface(color = MaterialTheme.colors.background) {
+                        HomePage()
+                    }
+                }
+            }
+        }
+    }
+}
