--- app/src/main/java/com/example/ihonboarding/presentation/ui/news/BaseApplication.kt
@@ -1,8 +1,0 @@
-package com.example.ihonboarding.presentation.ui.news
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class BaseApplication : Application() {
-}
